Isolate packages installation
Isolating the packages installation by similarity and estimated frequency of upgrades required per chunk.
As we use container images as a transport and although this could be considered an anti-pattern for the default container images use cases,
we'd prefer many small layers over a few big ones to decrease the probability of updating a single big layer each time we update.